feat: switch to vitest
