+++ test/wiiu-encrypt.test.ts
@@ -1,0 +1,22 @@
+import { describe, it, expect } from 'vitest';
+import { encryptWiiU } from '@pretendonetwork/boss-crypto';
+
+const BOSS_WIIU_AES_KEY = process.env.BOSS_WIIU_AES_KEY!;
+const BOSS_WIIU_HMAC_KEY = process.env.BOSS_WIIU_HMAC_KEY!;
+
+describe('encryptWiiU', () => {
+	it('should correctly encrypt Wii U BOSS data', () => {
+		const expected = Buffer.from([
+			0x62, 0x6F, 0x73, 0x73, 0x00, 0x02, 0x00, 0x01, 0x00, 0x01, 0x00, 0x02, 0x00, 0x00, 0x00, 0x00,
+			0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+			0xA8, 0x5E, 0xDB, 0xD7, 0x87, 0x41, 0x45, 0xFE, 0x47, 0x86, 0x19, 0x45, 0xF3, 0x94, 0x09, 0xE9,
+			0xD5, 0xE2, 0x81, 0xBB, 0x6C, 0xCD, 0xC7, 0xA6, 0x36, 0x40, 0x56, 0xCF, 0xA8, 0x41, 0x38, 0xF2,
+			0x51, 0x74, 0x7A, 0xDE, 0x1D, 0x78, 0x85, 0xC3, 0xD4, 0xA3, 0xA5
+		]);
+
+		const content = Buffer.from('Hello World');
+		const encrypted = encryptWiiU(content, BOSS_WIIU_AES_KEY, BOSS_WIIU_HMAC_KEY);
+
+		expect(encrypted.equals(expected)).toBe(true);
+	});
+});
