+++ vitest.config.ts
@@ -1,0 +1,8 @@
+import { loadEnv } from 'vite';
+import { defineConfig } from 'vitest/config';
+
+export default defineConfig(({ mode }) => ({
+	test: {
+		env: loadEnv(mode, process.cwd(), ''), // Loads all .env variables
+	},
+}))
